+++ packages/src/types/index.ts
@@ -1,3 +1,6 @@
 export * from './error';
 export * from './view';
 export * from './feature';
+export * from './lodash';
+
+export type SafeAny = any;